Initial working algorithm
This change implements a 2048 solver based entirely on a single metric: given action _a_ number of empty cells produced in 3-8 moves. Computational complexity grows exponentially with the number of look-ahead moves. Uncertainty from random tile generation is handled by making the same move multiple times. The exploration cuts off dead ends produced from illegal actions (e.g. nothing moves/combines) and losses. Therefore, if a particular move leads to high-probability losses within a few moves it will produce less output boards and therefore the total score of number of empty spaces in the resulting boards will be lower.

The code implements an in-memory model of 2048 a set of tests to exercise basic actions (though the test suite is incomplete). The entire logic of solving is in `Board.suggest` and a sample program that runs the game as far as possible can be found in `Main.scala`.